refactor(engine): split engine_sync.go and types.go by concern
engine_sync.go (950 lines) and types.go (890 lines) had become the two
biggest god-files left after P1. Pure mechanical file moves, zero behavior
change.

engine_sync.go (950 → 274 lines):
  - pull.go (98)         PullTrunk, ResetTrunkToRemote
  - restack_impl.go (594) restackBranch + the 4 private restack helpers
                          (the 300-line state machine that does the actual
                          per-branch work)
  - engine_sync.go (274)  RestackBranches family, ContinueRebase, Rebase
                          — the top-level orchestrators stay here.

types.go (890 → 382 lines):
  - branch.go (241)  Branch struct + branchReader interface + all of
                     Branch's methods
  - pr_info.go (278) PrInfo struct + constructors + getters + the With*
                     immutable-builder family
  - types.go (382)   enums (SortStrategy, LockReason, CommitFormat,
                     PullResult, RestackResult, RestackPlanAction,
                     ValidationResult, DeletionReasonKind), Scope,
                     status/result/option structs.

Phase F5 of the engine refactor round-2 runbook.
